+++ pyi.py
@@ -4,9 +4,11 @@ import logging
 
 import argparse
 import ast
+import re
 import sys
 from collections import Counter
 from collections.abc import Iterable, Sequence
+from copy import deepcopy
 from dataclasses import dataclass, field
 from flake8 import checker  # type: ignore
 from flake8.plugins.pyflakes import FlakesChecker  # type: ignore
@@ -425,6 +427,56 @@ class PyiVisitor(ast.NodeVisitor):
     def visit_AsyncFunctionDef(self, node: ast.AsyncFunctionDef) -> None:
         self._visit_function(node)
 
+    def _Y019_error(
+        self, node: ast.FunctionDef | ast.AsyncFunctionDef, typevar_name: str
+    ) -> None:
+        cleaned_method = deepcopy(node)
+        cleaned_method.decorator_list.clear()
+        new_syntax = re.sub(
+            fr"(\W){typevar_name}(\W)", r"\1Self\2", ast.unparse(cleaned_method)
+        ).replace("\n   ", "")
+        error_message = Y019.format(typevar_name=typevar_name, new_syntax=new_syntax)
+        self.error(node, error_message)
+
+    def _check_instance_method_for_bad_typevars(
+        self,
+        *,
+        method: ast.FunctionDef | ast.AsyncFunctionDef,
+        first_arg_annotation: ast.Name | ast.Subscript,
+        return_annotation: ast.Name,
+    ) -> None:
+        if not isinstance(first_arg_annotation, ast.Name):
+            return
+
+        if first_arg_annotation.id != return_annotation.id:
+            return
+
+        arg1_annotation_name = first_arg_annotation.id
+
+        if arg1_annotation_name.startswith("_"):
+            self._Y019_error(method, arg1_annotation_name)
+
+    def _check_class_method_for_bad_typevars(
+        self,
+        *,
+        method: ast.FunctionDef | ast.AsyncFunctionDef,
+        first_arg_annotation: ast.Name | ast.Subscript,
+        return_annotation: ast.Name,
+    ) -> None:
+        if not isinstance(first_arg_annotation, ast.Subscript):
+            return
+        if not isinstance(first_arg_annotation.slice, ast.Name):
+            return
+        if not isinstance(first_arg_annotation.value, ast.Name):
+            return
+        if first_arg_annotation.value.id != "type":
+            return
+
+        cls_typevar = first_arg_annotation.slice.id
+
+        if cls_typevar == return_annotation.id and cls_typevar.startswith("_"):
+            self._Y019_error(method, cls_typevar)
+
     def _visit_function(self, node: ast.FunctionDef | ast.AsyncFunctionDef) -> None:
         self._function_nesting += 1
         self.generic_visit(node)
@@ -441,6 +493,50 @@ class PyiVisitor(ast.NodeVisitor):
                 ):
                     continue
             self.error(statement, Y010)
+
+        # Rest of this method is devoted to finding TypeVars that should be _typeshed.Self, but aren't
+
+        if not self.in_class:
+            return
+        pos_or_keyword_args = node.args.args
+
+        if not pos_or_keyword_args:
+            return
+        return_annotation = node.returns
+
+        if not isinstance(return_annotation, ast.Name):
+            return
+        first_arg_annotation = pos_or_keyword_args[0].annotation
+
+        if not isinstance(first_arg_annotation, (ast.Name, ast.Subscript)):
+            return
+
+        if node.name == "__new__":
+            self._check_class_method_for_bad_typevars(
+                method=node,
+                first_arg_annotation=first_arg_annotation,
+                return_annotation=return_annotation,
+            )
+            return
+
+        for decorator in node.decorator_list:
+            if isinstance(decorator, ast.Name):
+                decorator_name = decorator.id
+                if decorator_name == "classmethod":
+                    self._check_class_method_for_bad_typevars(
+                        method=node,
+                        first_arg_annotation=first_arg_annotation,
+                        return_annotation=return_annotation,
+                    )
+                    return
+                elif decorator_name == "staticmethod":
+                    return
+
+        self._check_instance_method_for_bad_typevars(
+            method=node,
+            first_arg_annotation=first_arg_annotation,
+            return_annotation=return_annotation,
+        )
 
     def visit_arguments(self, node: ast.arguments) -> None:
         self.generic_visit(node)
@@ -560,6 +656,7 @@ Y015 = 'Y015 Attribute must not have a default value other than "..."'
 Y016 = "Y016 Duplicate union member"
 Y017 = "Y017 Only simple assignments allowed"
 Y018 = 'Y018 {typevarlike_cls} "{typevar_name}" is not used'
+Y019 = "Y019 Use `_typeshed.Self` instead of `{typevar_name}`, e.g. `{new_syntax}`"
 Y092 = "Y092 Top-level attribute must not have a default value"
 Y093 = "Y093 Use typing_extensions.TypeAlias for type aliases"