+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ unreleased
 * introduce Y017 (disallows assignments with multiple targets or non-name targets)
 * extend Y001 to cover ParamSpec and TypeVarTuple in addition to TypeVar
 * introduce Y018 (detect unused TypeVars)
+* introduce Y019 (detect TypeVars that should be _typeshed.Self, but aren't)
 * introduce Y016 (duplicate union member)
 * support Python 3.10
 * discontinue support for Python 3.6